requirement.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Image Augmentation
imgaug>=0.4.0

# Data Processing
numpy>=1.24.0
scipy>=1.10.0

# Geometry & Shapes
Shapely>=2.0.0
pyclipper>=1.3.0
Polygon3>=3.0.9

# Configuration & Utils
PyYAML>=6.0
anyconfig>=0.14.0
natsort>=8.3.0

# Visualization
matplotlib>=3.7.0
scikit-image>=0.21.0

# Progress & Logging
tqdm>=4.65.0
tensorboard>=2.13.0

# Additional
future>=0.18.3
addict>=2.4.0